requirements.txt
git+ssh://git@gitlab.tudelft.nl/qp4/netsquid-netrunner.git
git+ssh://git@gitlab.tudelft.nl/qp4/pyp4.git
git+ssh://git@gitlab.tudelft.nl/qp4/netsquid-p4.git
netsquid-physlayer
networkx
matplotlib
